+++ src/miner.cpp
@@ -15,7 +15,7 @@
 #include "program_util.hpp"
 
 const std::string Miner::ANONYMOUS("anonymous");
-const int64_t Miner::PROGRAMS_TO_FETCH = 2000;  // magic number
+const int64_t Miner::PROGRAMS_TO_FETCH = 5000;  // magic number
 const int64_t Miner::NUM_MUTATIONS = 100;       // magic number
 
 Miner::Miner(const Settings &settings, int64_t log_interval,